+++ dttools/src/b64_encode.h
@@ -7,8 +7,8 @@ See the file COPYING for details.
 #ifndef URL_DECODE_H
 #define URL_ENCODE_H
 
+#include <stddef.h>
 
-int b64_encode(const char *input, int len, char *output, int buf_len);
-
+int b64_encode(const char *input, size_t len, char *output, size_t buf_len);
 
 #endif
